pyproject.toml
[tool.poetry]
name = "pytrading"
version = "0.1.0"
description = ""
authors = ["Brice Boudehent <brice.boudehent@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pybit = "^2.1.1"
python-binance = "^1.0.16"
pandas = "^1.4.2"
mplfinance = "^0.12.8-beta.9"
Backtesting = "^0.3.3"
TA-Lib = "^0.4.24"
ccxt = "^1.81.72"
schedule = "^1.1.0"


[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
ipykernel = "^6.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
